added fallback in case cached achievement data is missing
